creo k ya

+++ panel/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-from django.contrib.auth.models import User 
+from django.contrib.auth.models import User
 from multiselectfield import MultiSelectField
 
 
@@ -15,7 +15,7 @@ class Producto(models.Model):
 
     usuario = models.ForeignKey(User, on_delete=models.CASCADE)  
     nombre = models.CharField(max_length=100)
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.IntegerField()  # Cambiado de DecimalField a IntegerField
     descripcion = models.TextField()
     imagen = models.ImageField(upload_to='productos/')
     activo = models.BooleanField(default=True)
